% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td_create_metadata_file.R
\name{td_create_metadata_file}
\alias{td_create_metadata_file}
\title{Create a new metadata template file}
\usage{
td_create_metadata_file(name, path = ".", format = "yaml", overwrite = FALSE)
}
\arguments{
\item{name}{a \code{character} of length 1. The trait dataset identifier used to
create files and folders. Should be short, explicit and without special
characters (including accents).}

\item{path}{a \code{character} of length 1. The folder name to stored the
metadata template file in. Must exist.
Default is the current directory.}

\item{format}{a \code{character} of length 1. One of \code{xlsx} or \code{yaml}.
Default is \code{yaml}.}

\item{overwrite}{a \code{logical} of length 1. If \code{TRUE} overwrites the metadata
template file.
Default is \code{FALSE}.}
}
\value{
No return value.
}
\description{
This function creates a new metadata template file used to describe a new
trait dataset. Two formats are currently available: \code{xlsx} or \code{yaml}. The
metadata template file will be stored in a subdirectory named \code{name}.
}
\examples{
\dontrun{
td_create_metadata_file(name = "pantheria")
}
}
