% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_yaml_xlsx.R
\name{yaml_as_df}
\alias{yaml_as_df}
\title{Transform a yaml metadata to data.frame}
\usage{
yaml_as_df(x)
}
\arguments{
\item{x}{a \code{character} of length 1. It could be a filename with extension '.yaml' or a full yaml string.}
}
\value{
A list with three data.frames: \code{status}, \code{dataset}, and \code{traits}.
}
\description{
This function transform a metadata written in yaml to
a metadata in data.frame format with three items:
\code{status}, \code{dataset}, and \code{traits}.
The structure of the metadata must follow the structure in \code{metadata_as_yaml()}
}
\examples{
\dontrun{
metadata_as_yaml() |> yaml_as_df()
}
}
