% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_yaml_xlsx.R
\name{yaml_to_xlsx}
\alias{yaml_to_xlsx}
\title{Create a xlsx file from a yaml metadata}
\usage{
yaml_to_xlsx(name, path = ".", out_suffix = "", overwrite = FALSE)
}
\arguments{
\item{name}{a \code{character} of length 1. The trait dataset identifier used to
create files and folders. Should be short, explicit and without special
characters (including accents).}

\item{path}{a \code{character} of length 1. The folder name to stored the
metadata template file in. Must exist.
Default is the current directory.}

\item{out_suffix}{a \code{character} of length 1. The suffix to be added to the output file.}

\item{overwite}{a \code{logical} of length 1. If \code{TRUE} overwrites the metadata
template file.
Default is \code{FALSE}.}
}
\value{
No return value.
}
\description{
This function transform a metadata saved in yaml format
into a metadata saved in a xlsx file.
}
